Change names and setup initial test
