+++ app/src/main/java/org/dhis2/data/dispatcher/DispatcherModule.kt
@@ -2,16 +2,16 @@ package org.dhis2.data.dispatcher
 
 import dagger.Module
 import dagger.Provides
+import javax.inject.Singleton
 import org.dhis2.form.model.DispatcherProvider
 import org.dhis2.form.model.coroutine.FormDispatcher
-import javax.inject.Singleton
 
 @Module
 open class DispatcherModule {
 
     @Provides
     @Singleton
-    open fun provideDispatcherModule(): DispatcherProvider{
+    open fun provideDispatcherModule(): DispatcherProvider {
         return FormDispatcher()
     }
 }
